% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggturtle.R
\name{ggturtle}
\alias{ggturtle}
\alias{turtle_init}
\alias{turtle_init,}
\alias{go_to,}
\alias{go_home,}
\alias{go_forward,}
\alias{go_back,}
\alias{pen_on,}
\alias{pen_off,}
\alias{turn_left,}
\alias{turn_right,}
\alias{set_size,}
\alias{set_color,}
\alias{set_alpha,}
\alias{turtle_draw}
\title{Use ggplot2 to draw like turtle in LOGO language.}
\usage{
turtle_init()
}
\description{
Use ggplot2 to draw like turtle in LOGO language.
}
\examples{
library(ggturtle)

a <- turtle_init(); for (i in 1:8) { a <- a \%>\% go_forward(100) \%>\% turn_right(135) }; a \%>\% turtle_draw("RP 8 [ FW 100, TR 135 ]")
}
